Some modifications to the classes, I still need to generate the Readme.md and the diagram
